// src/screenCast/types/ParamsStreamScrcpy.d.ts
import { ACTION } from '../common/Action';
import { ParamsStream } from './ParamsStream';
import VideoSettings from '../app/VideoSettings';

export interface ParamsStreamScrcpy extends ParamsStream {
  action: ACTION.STREAM_SCRCPY;
  ws: string;
  fitToScreen?: boolean;
  videoSettings?: VideoSettings;
  serial?: string;
  deviceName?: string;
}